+++ packages/network/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @0xsodium/network
 
+## 1.23.0
+
+### Minor Changes
+
+- fix wait user op
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/utils@1.23.0
+
 ## 1.22.0
 
 ### Minor Changes