+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @0xsodium/provider
 
+## 1.23.0
+
+### Minor Changes
+
+- fix wait user op
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/abi@1.23.0
+  - @0xsodium/auth@1.23.0
+  - @0xsodium/config@1.23.0
+  - @0xsodium/graphquery@1.23.0
+  - @0xsodium/network@1.23.0
+  - @0xsodium/transactions@1.23.0
+  - @0xsodium/utils@1.23.0
+  - @0xsodium/wallet@1.23.0
+
 ## 1.22.0
 
 ### Minor Changes