+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @0xsodium/utils
 
+## 1.23.0
+
+### Minor Changes
+
+- fix wait user op
+
 ## 1.22.0
 
 ### Minor Changes